feat: refactor searchable advocate table component to utilize heroui
- Renames component table-with-search to differentiate
- Utilizes chips for displaying specialties
- Adds asPhoneNumber function to format phone numbers for display
- Updates search logic to allow users to search for phone numbers with or without formatting characters
- Utilizes tailwind classes for styling
